iac-tagger/requirements.txt
# Core dependencies
pyyaml>=6.0.1
python-hcl2>=4.3.2
gitpython>=3.1.31

# Development dependencies
pytest>=7.4.0
black>=23.3.0
mypy>=1.4.1
flake8>=6.0.0